requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.11
djangorestframework==3.15.0
idna==3.6
PyJWT==2.8.0
requests==2.31.0
sqlparse==0.4.4
typing-extensions==4.10.0
urllib3==2.2.1
